split tools for analysis from processor
