+++ src/jobtrendx/analysis.py
@@ -30,7 +30,7 @@ import pandas as pd
 from omegaconf import DictConfig
 
 from . import logger
-from . import tools_processor as tools
+from . import tools_analysis as tools
 from . import payload_analysis
 from . import terms_unify
